% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugings_chico.R
\docType{data}
\name{gaugings_chico}
\alias{gaugings_chico}
\title{Gaugings at San Antonio Chico (direct stage discharge measurement)}
\format{
A data.frame with 2 columns:
\describe{
  \item{dates}{dates in POSIXct format}
  \item{h}{Water level (m)}
  \item{q}{discharge (m3/s)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
gaugings_chico
}
\description{
Gaugings at San Antonio Chico (direct stage discharge measurement)
}
\keyword{datasets}
