% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugings_ruta3.R
\docType{data}
\name{gaugings_ruta3}
\alias{gaugings_ruta3}
\title{Gaugings at Ruta 3 (direct stage discharge measurement)}
\format{
A data.frame with 2 columns:
\describe{
  \item{dates}{dates in POSIXct format}
  \item{h}{Water level (m)}
  \item{q}{discharge (m3/s)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
gaugings_ruta3
}
\description{
Gaugings at Ruta 3 (direct stage discharge measurement)
}
\keyword{datasets}
