% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec.R
\name{get_prec}
\alias{get_prec}
\title{Get precipitation from raingauge at a given timestep and period}
\usage{
get_prec(raingauge, timestep, ini.date, end.date)
}
\arguments{
\item{raingauge}{Character string that provides the code of the raingauge (i.e. "P01").}

\item{timestep}{Character string that provides the cumulative interval for precipitation depth. Can be “min”, “hour”, “day”, “DSTday”, “week”, “month”, “quarter” or “year”. (i.e. "10 min")}

\item{ini.date}{Object of class "POSIXlt with the start date}

\item{end.date}{Object of class "POSIXlt with the end date}
}
\value{
Tibble with two columns: date, and precipitation (mm)
}
\description{
Función para obtener la serie temporal de precipitación para una duración dada"
}
\examples{
get_prec(raingauge = "P01",
         timestep = "5 min",
         ini.date = strptime("2017-03-22 00:00:00", "\%Y-\%m-\%d \%H:\%M:\%S"),
         end.date = strptime("2017-03-23 00:00:00", "\%Y-\%m-\%d \%H:\%M:\%S"))
}
