% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_chico.R
\docType{data}
\name{level_chico}
\alias{level_chico}
\title{Raw water level at chico}
\format{
A data.frame with 2 columns:
\describe{
  \item{Date}{dates in POSIXct format}
  \item{LEVEL}{Water level (m)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
level_chico
}
\description{
Raw water level at chico
}
\keyword{datasets}
