% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_ruta3.R
\docType{data}
\name{level_ruta3}
\alias{level_ruta3}
\title{Raw water level at ruta3}
\format{
A data.frame with 2 columns:
\describe{
  \item{Date}{dates in POSIXct format}
  \item{LEVEL}{Water level (m)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
level_ruta3
}
\description{
Raw water level at ruta3
}
\keyword{datasets}
