% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P01.R
\docType{data}
\name{P01}
\alias{P01}
\title{Raw precipitation Padre Pio}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P01
}
\description{
Raw precipitation Padre Pio
}
\keyword{datasets}
