% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P02.R
\docType{data}
\name{P02}
\alias{P02}
\title{Raw precipitation Sodero}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P02
}
\description{
Raw precipitation Sodero
}
\keyword{datasets}
