% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P03.R
\docType{data}
\name{P03}
\alias{P03}
\title{Raw precipitation Gamorel}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P03
}
\description{
Raw precipitation Gamorel
}
\keyword{datasets}
