% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P04.R
\docType{data}
\name{P04}
\alias{P04}
\title{Raw precipitation Jorge Souza}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P04
}
\description{
Raw precipitation Jorge Souza
}
\keyword{datasets}
