% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P05.R
\docType{data}
\name{P05}
\alias{P05}
\title{Raw precipitation Caputto}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P05
}
\description{
Raw precipitation Caputto
}
\keyword{datasets}
