% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P06.R
\docType{data}
\name{P06}
\alias{P06}
\title{Raw precipitation Escuela 66 Tropezon}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P06
}
\description{
Raw precipitation Escuela 66 Tropezon
}
\keyword{datasets}
