% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P07.R
\docType{data}
\name{P07}
\alias{P07}
\title{Raw precipitation Abuelos}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P07
}
\description{
Raw precipitation Abuelos
}
\keyword{datasets}
