% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P08.R
\docType{data}
\name{P08}
\alias{P08}
\title{Raw precipitation Escuela 16 República Armenia}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P08
}
\description{
Raw precipitation Escuela 16 República Armenia
}
\keyword{datasets}
