% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P09.R
\docType{data}
\name{P09}
\alias{P09}
\title{Raw precipitation Escuela 69 Colonia Harriegue}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P09
}
\description{
Raw precipitation Escuela 69 Colonia Harriegue
}
\keyword{datasets}
