% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P10.R
\docType{data}
\name{P10}
\alias{P10}
\title{Raw precipitation Escuela 31 Colonia 18 de Julio}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P10
}
\description{
Raw precipitation Escuela 31 Colonia 18 de Julio
}
\keyword{datasets}
