% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P11.R
\docType{data}
\name{P11}
\alias{P11}
\title{Raw precipitation Invernadero}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P11
}
\description{
Raw precipitation Invernadero
}
\keyword{datasets}
