% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P12.R
\docType{data}
\name{P12}
\alias{P12}
\title{Raw precipitation Monetta}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P12
}
\description{
Raw precipitation Monetta
}
\keyword{datasets}
