% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P13.R
\docType{data}
\name{P13}
\alias{P13}
\title{Raw precipitation Saracho}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P13
}
\description{
Raw precipitation Saracho
}
\keyword{datasets}
