% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P14.R
\docType{data}
\name{P14}
\alias{P14}
\title{Raw precipitation Uruguay}
\format{
A data.frame with 2 columns:
\describe{
  \item{date}{dates in POSIXct format}
  \item{precipitation}{Precipitation depth (mm)}
}
}
\source{
\url{http://agua.unorte.edu.uy/}
}
\usage{
P14
}
\description{
Raw precipitation Uruguay
}
\keyword{datasets}
