% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SanAntonioData-package.R
\docType{package}
\name{SanAntonioData-package}
\alias{SanAntonioData}
\alias{SanAntonioData-package}
\title{SanAntonioData: Hydrometeorological dataset of San Antonio Basin - Uruguay}
\description{
Water level, gaugings, precipitation, streamflows datasets of San Antonio Basin - Uruguay.
}
\author{
\strong{Maintainer}: Rafael Navas \email{rafaelnavas23@gmail.com} (\href{https://orcid.org/0000-0001-8559-9523}{ORCID})

}
\keyword{internal}
